DS-24 move modules to 'src', use data-attrs for click-handler
