+++ packages/core-v3.x/modules/src/click-handler.js
@@ -11,8 +11,8 @@ export default class ClickHandler {
     // where props are automatically turned into node data. We're adapting it
     // to work with non-WCs by manually setting that data from attributes.
 
-    this.el.onClick = this.el.getAttribute('on-click');
-    this.el.onClickTarget = this.el.getAttribute('on-click-target');
+    this.el.onClick = this.el.dataset.onClick;
+    this.el.onClickTarget = this.el.dataset.onClickTarget;
 
     this.el.addEventListener('click', () => {
       declarativeClickHandler(this.el);
